Reversed the logic for set_square.
When using if-else clauses, 'is' should probably be the first condition, unless it's an unusual situation. Even then, should 'is' be the first condition for consistency in general cases? Or should the most common situation be the first?

I don't know that answer, but regardless, empty tiles are statistically more common on a per-move basis. So I might have been thinking backwards at the time I wrote that code.